Create new file for Analytics data. Relates to #69.
